Rename Android component to org.rustls.platformverifier
This is more consistent with other Android libraries
