+++ android/rustls-platform-verifier/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="rustls.platformverifier.android"/>
+    package="org.rustls.platformverifier"/>
