+++ android/rustls-platform-verifier/src/main/java/org/rustls/platformverifier/CertificateVerifier.kt
@@ -1,4 +1,4 @@
-package rustls.platformverifier.android
+package org.rustls.platformverifier
 
 import android.content.Context
 import android.net.http.X509TrustManagerExtensions
